CRUD of clients and services ready

+++ Models/Servicio.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace EmpresaTVxCable.Models;
 
@@ -9,7 +10,7 @@ public partial class Servicio
 
     public string Nombre { get; set; } = null!;
 
-    public decimal Precio { get; set; }
+    public int Precio { get; set; }
 
     public virtual ICollection<Contrato> Contratos { get; set; } = new List<Contrato>();
 }
